+++ viewWidget.py
@@ -13,11 +13,11 @@ class viewWidget(QGroupBox):
 
     show_all = pyqtSignal()
     zoom_by_increment = pyqtSignal(str, int)
-    scale_font_size_changed = pyqtSignal(float)
-    label_font_size_changed = pyqtSignal(float)
+    scale_font_size_changed = pyqtSignal(np.float64)
+    label_font_size_changed = pyqtSignal(np.float64)
     fig_size_changed = pyqtSignal(tuple)
     annotation_changed = pyqtSignal(str)
-    annotation_font_size_changed = pyqtSignal(float)
+    annotation_font_size_changed = pyqtSignal(np.float64)
     resizing_changed = pyqtSignal(bool)
 
     def __init__(self):
@@ -238,15 +238,15 @@ class viewWidget(QGroupBox):
                         if (item[0] == 'atx'):  # annotation text
                             self.set_annotation_text(str(item[1]).replace('\\n', '\n'))
                         elif (item[0] == 'afs'):  # annotation font size
-                            self.set_annotation_font_size(float(item[1]))
+                            self.set_annotation_font_size(np.float64(item[1]))
                         elif (item[0] == 'fiw'):  # fig size width
-                            self.set_fig_size((float(item[1]), None))
+                            self.set_fig_size((np.float64(item[1]), None))
                         elif (item[0] == 'fih'):  # fig size height
-                            self.set_fig_size((None, float(item[1])))
+                            self.set_fig_size((None, np.float64(item[1])))
                         elif (item[0] == 'lfs'):  # label font size
-                            self.set_label_font_size(float(item[1]))
+                            self.set_label_font_size(np.float64(item[1]))
                         elif (item[0] == 'sfs'):  # scale font size
-                            self.set_scale_font_size(float(item[1]))
+                            self.set_scale_font_size(np.float64(item[1]))
                         elif (item[0] == 'fse'):  # fig size enabled
                             if item[0] == '1' or item[0] == 'True':
                                 self.set_fig_size_enabled(True)
